+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - popup: Dashboard was not active on non-bandcamp pages
 - popup: Console was broken
+- popup: Fixed release Discogs search link
 
 ## 0.11.0 (2024-07-29)